v3: revamp: address review comments
